third_party/blink/web_tests/platform/generic/touchadjustment/touch-links-active-expected.txt

Tests that tapDown will trigger the active state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing indirect hits.
PASS isLinkActive() is true
PASS isLinkActive() is true
PASS isLinkActive() is true
Testing direct hits.
PASS isLinkActive() is true
PASS isLinkActive() is true
PASS isLinkActive() is true
PASS successfullyParsed is true

TEST COMPLETE
